部门职务

+++ src/main/java/com/stylefeng/guns/common/constant/Const.java
@@ -21,7 +21,7 @@ public interface Const {
     /**
      * 管理员id
      */
-    Integer ADMIN_ID = 1;
+    Long ADMIN_ID = 1L;
 
     /**
      * 超级管理员角色id
